# Batch processing in process_add_dir and add 1m files benchmark