Add Performance Benchmark:
| Test Case | Before (ms) | After (ms) | Change |
|-----------|-----------|----------------|-------------|
| 1k_20dir | 8.172 | 7.494 | -8.30% |
| 10k_20dir | 42.062 | 41.435 | -1.49% |
| 100k_20dir | 390.98 | 384.54 | -1.65% |
| 100k_100dir | 380.15 | 377.41 | -0.72% |
| 1m_1000dir | 9297.1 | 9186 | -1.19% |

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **New Features**
  * Enhanced benchmarking capabilities with support for varying repository and directory sizes, allowing more flexible performance measurements.

* **Performance Improvements**
  * Improved the efficiency of adding directories by batching file processing, which may result in faster operations when working with large directories.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->