@Lilianxr 2025-07-25
Update benchmark performance result:
| Benchmark       | Main (ms) | Async-add (ms) | Change    |
|-----------------|-----------|----------------|-----------|
| 1k_20dir        | 7.63      | 6.9038         | -9.54%    |
| 10k_20dir       | 42.15     | 41.059         | -2.58%    |
| 100k_20dir      | 392.33    | 384.12         | -2.09%    |
| 100k_100dir     | 377.49    | 374.6          | -0.76%    |
| 100k_1000dir    | 435.80    | 432.18         | -0.83%    |
| 1m_1000dir      | 9297.47   | 9046.8         | -2.70%    |

Tested with batch size of 16, 32, 64, 128, and concurrency of cpu_num * 2 and cpu_num * 3. 
The best param set with the above repo structures is batch size of 64 and concurrency of cpu_num * 2